+++ src/core/constants.js
@@ -1,0 +1,5 @@
+const CONSTANTS = {
+    NEXO_VERSION: "2.1.0"
+}
+
+export default CONSTANTS;