+++ src/lib/file_class.js
@@ -1,0 +1,12 @@
+class Correspond{
+    from;
+    to;
+    constructor(from,to){
+        this.from = from;
+        this.to = to;
+    }
+}
+
+export {
+    Correspond
+}